move move message processing logic out of CWsConnection and into (new) CMessageProcessor)
